convert tests (#1559)
